+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.0.2]
+
+* Breaking Change: order related parameters moved to QueryOrder object.
+* unnecessary dependencies removed
+* lazy-load bug fix when display compare function provided
+* better log
+
 ## [0.0.1]
 
 * initialization